requirements.txt
discord.py>=1.7.2
aiosqlite>=0.17.0
aPRAW>=0.6.8a0